person_server/requirements.txt
asgiref==3.3.0
attrs==19.3.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
curlify==2.2.1
Django==3.1.3
django-stubs==1.7.0
djangorestframework==3.12.2
djangorestframework-stubs==1.3.0
drf-spectacular==0.11.0
graphql-core==3.1.2
hypothesis==5.41.2
hypothesis-graphql==0.3.2
hypothesis-jsonschema==0.18.0
idna==2.10
inflection==0.5.1
iniconfig==1.1.1
isodate==0.6.0
itypes==1.2.0
Jinja2==2.11.2
jsonschema==3.2.0
junit-xml==1.9
lazy-object-proxy==1.5.1
MarkupSafe==1.1.1
more-itertools==8.6.0
multidict==5.0.0
mypy==0.790
mypy-extensions==0.4.3
openapi-core==0.13.4
openapi-schema-validator==0.1.1
openapi-spec-validator==0.2.9
packaging==20.4
parse==1.18.0
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.1.2
pytest-subtests==0.3.2
pytz==2020.4
PyYAML==5.3.1
requests==2.24.0
schemathesis==2.7.5
six==1.15.0
sortedcontainers==2.3.0
sqlparse==0.4.1
starlette==0.13.8
strict-rfc3339==0.7
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.11
Werkzeug==1.0.1
yarl==1.6.2
